refactor: reorganise components into views/, add #/ import alias
- Extract Header, MainView into dedicated files; rename SettingsPanel → SettingsView
- Move views into src/views/ directory
- Replace boolean settingsOpen with typed View = "main" | "settings" state
- Add package.json imports field and tsconfig paths for #/* → src/* alias
